Refactor Bosche structs

+++ Core/Inc/common.h
@@ -255,24 +255,24 @@ typedef struct {
 } BMx680_calib_t;
 
 typedef struct {
-  int16_t dig_t1;
-  int16_t dig_t2;
-  int16_t dig_t3;
-  int16_t dig_p1;
-  int16_t dig_p2;
-  int16_t dig_p3;
-  int16_t dig_p4;
-  int16_t dig_p5;
-  int16_t dig_p6;
-  int16_t dig_p7;
-  int16_t dig_p8;
-  int16_t dig_p9;
-  int16_t dig_h1;
-  int16_t dig_h2;
-  int16_t dig_h3;
-  int16_t dig_h4;
-  int16_t dig_h5;
-  int16_t dig_h6;
+  uint16_t  dig_t1;
+  int16_t   dig_t2;
+  int16_t   dig_t3;
+  uint16_t  dig_p1;
+  int16_t   dig_p2;
+  int16_t   dig_p3;
+  int16_t   dig_p4;
+  int16_t   dig_p5;
+  int16_t   dig_p6;
+  int16_t   dig_p7;
+  int16_t   dig_p8;
+  int16_t   dig_p9;
+  uint8_t   dig_h1;
+  int16_t   dig_h2;
+  uint8_t   dig_h3;
+  int16_t   dig_h4;
+  int16_t   dig_h5;
+  int8_t    dig_h6;
 } BMx280_calib_t;
 
 /**
